Feat & Fix: Functions
We finished creating the menu function and along the way we made many corrections to the code such as search functions that had recusive bugs, then functions to present information in the console, and functions to add code to the vectors.

The set of these fixes make the program works quite well, although it still has bugs that remain to be corrected.